nauka
zmiana nazw plików zwiększająca przejrzystosc
