--- django_css/requirements.txt
@@ -1,3 +1,0 @@
-Django==2.2.6
-pandas
-scipy
