--- django_css/uploads/core/forms.py
@@ -1,9 +1,0 @@
-from django import forms
-
-from uploads.core.models import Document
-
-
-class DocumentForm(forms.ModelForm):
-    class Meta:
-        model = Document
-        fields = ('description', 'document', )
