--- django_css/uploads/core/models.py
@@ -1,3 +1,0 @@
-from __future__ import unicode_literals
-
-from django.db import models
