--- django_css/uploads/templates/about.html
@@ -1,7 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-    <h1>About</h1>
-
-    <p>Example information</p>
-{% endblock %}
