--- django_css/uploads/templates/base.html
@@ -1,23 +1,0 @@
-{% load static %}
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <title>Python dla początkujących</title>
-    <link rel="stylesheet" href="{% static 'css/main.css' %}">
-  </head>
-  <body>
-    <header>
-      <nav>
-          <ul>
-              <li><a href="{% url 'home' %}">Home</a></li>
-              <li><a href="{% url 'data_analysis' %}">Data analysis</a></li>
-              <li><a href="{% url 'about' %}">About</a></li>
-              <li><a href="{% url 'contact' %}">Contact</a></li>
-          </ul>
-      </nav>
-    </header>
-    {% block content %}
-    {% endblock %}
-  </body>
-</html>
