--- django_css/uploads/templates/contact.html
@@ -1,7 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-    <h1>Contact</h1>
-
-    <p>How to contact me</p>
-{% endblock %}
