--- django_css/uploads/templates/data_analysis.html
@@ -1,29 +1,0 @@
-{% extends 'base.html' %}
-
-{% load static %}
-
-{% block content %}
-  <p>
-    <form method="post" enctype="multipart/form-data">
-      {% csrf_token %}
-      <input type="file" name="myfile">
-      <input type="submit" value="Upload and analyze">
-    </form>
-  </p>
-
-  {% if result_present %}
-    <p><strong>Result:</strong>
-      <ul>
-        <li>Preason's r<br>
-            {{ results.r_table|safe }}</li>
-        <li>probability <i>p</i> < x <br>
-            {{ results.p_table|safe }}</li>
-      </ul>
-    </p>
-
-      <p>Raw data, pandas' DataFrame rendered as HTML</p>
-      <p>{{ df|safe }}</p>
-
-  {% endif %}
-
-{% endblock %}
