+++ nauka/requirements.txt
@@ -1,0 +1,3 @@
+Django==2.2.6
+pandas
+scipy
