+++ nauka/uploads/core/forms.py
@@ -1,0 +1,9 @@
+from django import forms
+
+from uploads.core.models import Document
+
+
+class DocumentForm(forms.ModelForm):
+    class Meta:
+        model = Document
+        fields = ('description', 'document', )
