+++ nauka/uploads/templates/about.html
@@ -1,0 +1,7 @@
+{% extends 'base.html' %}
+
+{% block content %}
+    <h1>About</h1>
+
+    <p>Example information</p>
+{% endblock %}
