+++ nauka/uploads/templates/contact.html
@@ -1,0 +1,7 @@
+{% extends 'base.html' %}
+
+{% block content %}
+    <h1>Contact</h1>
+
+    <p>How to contact me</p>
+{% endblock %}
